networking: Assume host networking when the hostname resolves to any single IP address
Checking specifically for '172.x.x.x' doesn't work if the Docker network
is configured to use a different range, or with Podman (which uses
10.0.0.x).

It should be fairly rare to have a single IP on the host, so this should
mostly work.

The code was also refactored a bunch, with no other changes in its
functionality.